[UPDATE] update new casl guard for order, product, shop

+++ src/modules/product/services/product.service.ts
@@ -1,10 +1,17 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  ForbiddenException,
+  Injectable,
+} from '@nestjs/common';
 import { ProductRepository } from '../repositories/product.repository';
 import { ProductCreateDto } from '../dtos/product.create.dto';
 import { ProductUpdateDto } from '../dtos/product.update.dto';
 import { Product } from '../entities/product.entity';
 import { ProductChangeStatusDto } from '../dtos/product.change-status.dto';
 import { FindOneOptions, FindOptions, FindOptionsWhere } from 'typeorm';
+import { AppAbility } from 'src/modules/casl/casl-ability.factory';
+import { Action } from 'src/modules/casl/constants/casl.constant';
+import { subject } from '@casl/ability';
 
 @Injectable()
 export class ProductService {
@@ -19,7 +26,7 @@ export class ProductService {
   }
 
   async findById(id: number): Promise<Product> {
-    return await this.productRepository.findOneById(id);
+    return await this.productRepository.findById(id);
   }
 
   async createProduct(productCreateDto: ProductCreateDto): Promise<Product> {
@@ -29,18 +36,43 @@ export class ProductService {
   async updateProduct(
     id: number,
     productUpdateDto: ProductUpdateDto,
+    abilities: AppAbility,
   ): Promise<Product> {
+    const product = await this.findById(id);
+
+    if (!abilities?.can(Action.Update, subject('Product', product))) {
+      throw new ForbiddenException('Your shop can not update this resource');
+    }
+
     return await this.productRepository.updateProduct(id, productUpdateDto);
   }
 
-  async deleteProduct(id: number): Promise<void> {
+  async deleteProduct(id: number, abilities: AppAbility): Promise<void> {
+    const product = await this.findById(id);
+
+    if (!abilities?.can(Action.Delete, subject('Product', product))) {
+      throw new ForbiddenException('Your shop can not delete this resource');
+    }
+
     await this.productRepository.delete(id);
   }
 
   async changeStatusProduct(
     id: number,
     productchangeStatusDto: ProductChangeStatusDto,
+    abilities: AppAbility,
   ): Promise<Product> {
+    const { isAlready } = productchangeStatusDto;
+    const product = await this.findById(id);
+
+    if (!abilities?.can(Action.Update, subject('Product', product))) {
+      throw new ForbiddenException('Your shop can not update this resource');
+    }
+
+    if (product.isAlready === isAlready) {
+      throw new BadRequestException('New status no different');
+    }
+
     return await this.productRepository.changeStatusProduct(
       id,
       productchangeStatusDto,
